requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
annotated-types==0.7.0
    # via
    #   -r requirements.in
    #   pydantic
anyio==4.9.0
    # via
    #   -r requirements.in
    #   starlette
blis==1.3.0
    # via
    #   -r requirements.in
    #   thinc
catalogue==2.0.10
    # via
    #   -r requirements.in
    #   spacy
    #   srsly
    #   thinc
certifi==2025.4.26
    # via
    #   -r requirements.in
    #   requests
charset-normalizer==3.4.2
    # via
    #   -r requirements.in
    #   requests
click==8.2.0
    # via
    #   -r requirements.in
    #   typer
    #   uvicorn
cloudpathlib==0.21.1
    # via
    #   -r requirements.in
    #   weasel
confection==0.1.5
    # via
    #   -r requirements.in
    #   thinc
    #   weasel
cymem==2.0.11
    # via
    #   -r requirements.in
    #   preshed
    #   spacy
    #   thinc
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
    # via -r requirements.in
exceptiongroup==1.3.0
    # via
    #   -r requirements.in
    #   anyio
fastapi==0.115.12
    # via -r requirements.in
filelock==3.18.0
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2025.3.2
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   torch
h11==0.16.0
    # via
    #   -r requirements.in
    #   uvicorn
huggingface-hub==0.31.2
    # via
    #   -r requirements.in
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   -r requirements.in
    #   anyio
    #   requests
jinja2==3.1.6
    # via
    #   -r requirements.in
    #   spacy
    #   torch
langcodes==3.5.0
    # via
    #   -r requirements.in
    #   spacy
language-data==1.3.0
    # via
    #   -r requirements.in
    #   langcodes
marisa-trie==1.2.1
    # via
    #   -r requirements.in
    #   language-data
markdown-it-py==3.0.0
    # via
    #   -r requirements.in
    #   rich
markupsafe==3.0.2
    # via
    #   -r requirements.in
    #   jinja2
mdurl==0.1.2
    # via
    #   -r requirements.in
    #   markdown-it-py
mpmath==1.3.0
    # via
    #   -r requirements.in
    #   sympy
murmurhash==1.0.12
    # via
    #   -r requirements.in
    #   preshed
    #   spacy
    #   thinc
networkx==3.4.2
    # via
    #   -r requirements.in
    #   torch
numpy==2.2.5
    # via
    #   -r requirements.in
    #   blis
    #   spacy
    #   spacy-transformers
    #   thinc
    #   transformers
nvidia-cublas-cu12==12.6.4.1
    # via
    #   -r requirements.in
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.6.80
    # via
    #   -r requirements.in
    #   torch
nvidia-cuda-nvrtc-cu12==12.6.77
    # via
    #   -r requirements.in
    #   torch
nvidia-cuda-runtime-cu12==12.6.77
    # via
    #   -r requirements.in
    #   torch
nvidia-cudnn-cu12==9.5.1.17
    # via
    #   -r requirements.in
    #   torch
nvidia-cufft-cu12==11.3.0.4
    # via
    #   -r requirements.in
    #   torch
nvidia-cufile-cu12==1.11.1.6
    # via
    #   -r requirements.in
    #   torch
nvidia-curand-cu12==10.3.7.77
    # via
    #   -r requirements.in
    #   torch
nvidia-cusolver-cu12==11.7.1.2
    # via
    #   -r requirements.in
    #   torch
nvidia-cusparse-cu12==12.5.4.2
    # via
    #   -r requirements.in
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.6.3
    # via
    #   -r requirements.in
    #   torch
nvidia-nccl-cu12==2.26.2
    # via
    #   -r requirements.in
    #   torch
nvidia-nvjitlink-cu12==12.6.85
    # via
    #   -r requirements.in
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.6.77
    # via
    #   -r requirements.in
    #   torch
packaging==25.0
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   spacy
    #   thinc
    #   transformers
    #   weasel
preshed==3.0.9
    # via
    #   -r requirements.in
    #   spacy
    #   thinc
pydantic==2.11.4
    # via
    #   -r requirements.in
    #   confection
    #   fastapi
    #   spacy
    #   thinc
    #   weasel
pydantic-core==2.33.2
    # via
    #   -r requirements.in
    #   pydantic
pygments==2.19.1
    # via
    #   -r requirements.in
    #   rich
python-dotenv==1.1.0
    # via -r requirements.in
python-multipart==0.0.20
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   transformers
regex==2024.11.6
    # via
    #   -r requirements.in
    #   transformers
requests==2.32.3
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   spacy
    #   transformers
    #   weasel
rich==14.0.0
    # via
    #   -r requirements.in
    #   typer
safetensors==0.5.3
    # via
    #   -r requirements.in
    #   transformers
sentencepiece==0.2.0
    # via -r requirements.in
setuptools==80.7.1
    # via
    #   -r requirements.in
    #   marisa-trie
    #   spacy
    #   thinc
    #   triton
shellingham==1.5.4
    # via
    #   -r requirements.in
    #   typer
smart-open==7.1.0
    # via
    #   -r requirements.in
    #   weasel
sniffio==1.3.1
    # via
    #   -r requirements.in
    #   anyio
spacy==3.8.5
    # via
    #   -r requirements.in
    #   spacy-transformers
spacy-alignments==0.9.1
    # via
    #   -r requirements.in
    #   spacy-transformers
spacy-legacy==3.0.12
    # via
    #   -r requirements.in
    #   spacy
spacy-loggers==1.0.5
    # via
    #   -r requirements.in
    #   spacy
spacy-transformers==1.3.8
    # via -r requirements.in
srsly==2.5.1
    # via
    #   -r requirements.in
    #   confection
    #   spacy
    #   spacy-transformers
    #   thinc
    #   weasel
starlette==0.46.2
    # via
    #   -r requirements.in
    #   fastapi
sympy==1.14.0
    # via
    #   -r requirements.in
    #   torch
thinc==8.3.6
    # via
    #   -r requirements.in
    #   spacy
tokenizers==0.21.1
    # via
    #   -r requirements.in
    #   transformers
torch==2.7.0
    # via
    #   -r requirements.in
    #   spacy-transformers
tqdm==4.67.1
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   spacy
    #   transformers
transformers==4.49.0
    # via
    #   -r requirements.in
    #   spacy-transformers
triton==3.3.0
    # via
    #   -r requirements.in
    #   torch
typer==0.15.3
    # via
    #   -r requirements.in
    #   spacy
    #   weasel
typing-extensions==4.13.2
    # via
    #   -r requirements.in
    #   anyio
    #   cloudpathlib
    #   exceptiongroup
    #   fastapi
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   rich
    #   torch
    #   typer
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.0
    # via
    #   -r requirements.in
    #   pydantic
urllib3==2.4.0
    # via
    #   -r requirements.in
    #   requests
uvicorn==0.34.2
    # via -r requirements.in
wasabi==1.1.3
    # via
    #   -r requirements.in
    #   spacy
    #   thinc
    #   weasel
weasel==0.4.1
    # via
    #   -r requirements.in
    #   spacy
wrapt==1.17.2
    # via
    #   -r requirements.in
    #   smart-open
xx-ent-wiki-sm @ https://github.com/explosion/spacy-models/releases/download/xx_ent_wiki_sm-3.8.0/xx_ent_wiki_sm-3.8.0-py3-none-any.whl
    # via -r requirements.in
